start major refactor
