+++ Project.toml
@@ -18,7 +18,6 @@ PeriodicTable = "7b2266bf-644c-5ea3-82d8-af4bbd25a884"
 PrecompileSignatures = "91cefc8d-f054-46dc-8f8c-26e11d7c5411"
 RDKitMinimalLib = "44044271-7623-48dc-8250-42433c44e4b7"
 SharedArrays = "1a1011a3-84de-559e-8e89-a11a2f7dc383"
-SparseArrays = "2f01184e-e22b-5df5-ae63-d93ebab69eaf"
 
 [compat]
 Aqua = "^0.5"
