+++ docs/src/product_graphs.md
@@ -22,11 +22,11 @@ subtypes(AbstractProductGraph)
 ```
 
 While a [`ProductGraph`](@ref) contains the full representation of the product graph between two molecules, the adjacency matrix may be all that is required.
-As such, it may be more efficient to work with the adjacency matrix (see [`product_graph_adjacency_matrix`](@ref)).
+As such, it may be more efficient to work with the adjacency matrix (see [`GraphMatrix`](@ref)).
 
 To create a new [`ProductGraph`](@ref), define a new `AbstractProductGraph` subtype and two functions, `record_adjacency!` and `product_graph_edge_label`, parameterized by the new type.
 
 ```@docs
 ProductGraph
-product_graph_adjacency_matrix
+GraphMatrix
 ```
