+++ src/graph_products/weighted.jl
@@ -1,0 +1,1 @@
+export Weighted
